flask/requirements.txt
Click==8.1.3
Flask==2.2.2
mysql-connector-python
flask-cors
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
selenium==4.27.1
